+++ packages/atom-react-redux/__tests__/store/withCallback.test.ts
@@ -1,6 +1,4 @@
-import { put } from "redux-saga/effects";
-import * as common from "@/common";
-import { withCallback, withLoading } from "@/utils";
+import { withCallback } from "@/withCallback";
 
 function* successfulNumberSaga(): Generator<unknown, number, unknown> {
     return 55;
@@ -14,39 +12,6 @@ function* pausedSaga(): Generator<string, number, unknown> {
     yield "paused";
     return 1;
 }
-
-describe("withLoading", () => {
-    test("Given successful saga Then dispatches show and hide around execution", () => {
-        // Arrange
-        const loadingTarget = 7;
-        const iterator = withLoading(loadingTarget, successfulNumberSaga);
-
-        // Act
-        const first = iterator.next();
-        const second = iterator.next();
-        const third = iterator.next();
-
-        // Verify
-        expect(first.value).toEqual(put(common.Actions.showLoader(loadingTarget)));
-        expect(second.value).toEqual(put(common.Actions.hideLoader(loadingTarget)));
-        expect(third).toEqual({ value: 55, done: true });
-    });
-
-    test("Given failing saga Then still dispatches hide in finally and rethrows", () => {
-        // Arrange
-        const loadingTarget = 9;
-        const iterator = withLoading(loadingTarget, throwingSaga);
-
-        // Act
-        const first = iterator.next();
-        const second = iterator.next();
-
-        // Verify
-        expect(first.value).toEqual(put(common.Actions.showLoader(loadingTarget)));
-        expect(second.value).toEqual(put(common.Actions.hideLoader(loadingTarget)));
-        expect(() => iterator.next()).toThrow("Saga failed");
-    });
-});
 
 describe("withCallback", () => {
     test("Given successful saga and resolve callback Then resolves with result", () => {