+++ packages/atom-react-redux/src/common/sagas.ts
@@ -1,9 +1,7 @@
 import { call } from "redux-saga/effects";
 import { AxiosError } from "axios";
-import { withComponentName } from "@hwndmaster/atom-web-core";
+import { withComponentName, isDev, getNotificationService } from "@hwndmaster/atom-web-core";
 import type { ErrorInfo, RaiseErrorInfo, ValidationError } from "@hwndmaster/atom-web-core";
-import { isDev } from "@hwndmaster/atom-web-core";
-import { getNotificationService } from "@/notifications";
 import type { SagaGenerator } from "@/types";
 import * as commonActions from "./actions";