+++ packages/atom-react-prime/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hwndmaster/atom-react-prime
 
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @hwndmaster/atom-web-core@0.1.5
+  - @hwndmaster/atom-react-core@0.1.9
+
 ## 0.1.12
 
 ### Patch Changes